Refactor: Consolidate duplicate EntryPoints
Consolidate the duplicated `NetworkDependenciesEntryPoint` into `ServiceDependenciesEntryPoint`.
* Merged `apiInterface`, `syncManager`, `uploadManager`, `uploadToShelfService`, and `retryQueue` into `ServiceDependenciesEntryPoint`.
* Updated usages in `MainApplication.kt` to depend on the consolidated entry point.
* Removed obsolete `NetworkDependenciesEntryPoint.kt`.
